:truck: Move and rename files to keep order
